option to not create table and just create MessageTable object for querying and plotting purposes

+++ src/MessageTable.py
@@ -1,15 +1,16 @@
 from MessageDatabase import REQUIRED_PARAMETERS
 from Table import Table
-from Utils.General import unix_time_to_date_time_string, are_required_parameters_in_message, create_value_formatters
+from Utils.General import unix_time_to_date_time_string, are_required_parameters_in_message
 
-REQUIRED_PARAMETERS = ['timestamp', 'sourceUuid', 'body']
+REQUIRED_PARAMETERS = ['timestamp', 'sourceUuid', 'body', 'conversationId']
 
 class MessageTable(Table):
-    def __init__(self, cursor, conversation_id):
-        super().__init__(cursor, conversation_id)
+    def __init__(self, cursor, conversation_id, create_new = True):
+        super().__init__(cursor, conversation_id, create_new)
         self.table_name = "Messages" + self.conversation_id.replace("-", "")
-        self.cursor.execute("DROP TABLE IF EXISTS %s" % self.table_name)
-        self.cursor.execute("CREATE TABLE %s (timestamp text, source_uuid TEXT, body TEXT)" % self.table_name)
+        if create_new:
+            self.cursor.execute("DROP TABLE IF EXISTS %s" % self.table_name)
+            self.cursor.execute("CREATE TABLE %s (timestamp text, source_uuid TEXT, body TEXT)" % self.table_name)
 
     def load_table(self, message=""):
         if are_required_parameters_in_message(message, REQUIRED_PARAMETERS) and message['conversationId'] == self.conversation_id:
